+++ src/macro-action-streaming.cpp
@@ -43,7 +43,7 @@ void MacroActionStream::LogAction()
 		vblog(LOG_INFO, "performed action \"%s\"", it->second.c_str());
 	} else {
 		blog(LOG_WARNING, "ignored unknown streaming action %d",
-		     _action);
+		     static_cast<int>(_action));
 	}
 }
